added the option to pass normal find params to ff( findfile ) with --params
added copy to make file
